fix: tooltip hidden with custom checkbox & fix typo on custom checkbox
